refactor: WIP - ActionButton structure
