+++ Source/ClickerGame/Private/Systems/Economy/Data/EconomySnapshot.cpp
@@ -1,2 +1,2 @@
-#include "Data/Economy/EconomySnapshot.h"
+#include "Systems/Economy/Data/EconomySnapshot.h"
