+++ Source/ClickerGame/Private/Systems/UI/Data/ActionButton/ActionButtonDefinition.cpp
@@ -1,4 +1,4 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "Systems/UI/Widgets/Action/ActionButtonDefinition.h"
+#include "Systems/UI/Data/ActionButton/ActionButtonDefinition.h"
