+++ Source/ClickerGame/Private/Systems/UI/Types/ActionButtonTypes.cpp
@@ -1,5 +1,5 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "Systems/UI/ActionButtonTypes.h"
+#include "Systems/UI/Types/ActionButtonTypes.h"
