+++ Source/ClickerGame/Public/Systems/Save/Data/ClickerSaveGame.h
@@ -4,6 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/SaveGame.h"
+
 #include "ClickerSaveGame.generated.h"
 
 /**
