+++ docs/planning/2026/Daily/2026-02/2026-02-02.md
@@ -1,12 +1,11 @@
-﻿# 2026-01-30
+﻿# 2026-02-02
 ## Topic
 
 - WBP 구조 리팩터
 
 ## Tasks
 
-- [x] WBP 구조 리팩터
-  - [x] WIP - ActionButton WBP, C++ 구현
+- [x] WBP 구조 리팩터  
   - [x] 파일 구조 정리
 
 ## Blockers
@@ -25,10 +24,10 @@
 ## EOD
 
 - Done:
-  - WBP 구조 리팩터
-    - WIP - ActionButton WBP, C++ 구현
+  - WBP 구조 리팩터  
     - 파일 구조 정리
 - Evidence:
   - Commit Link
 - Carry to Tomorrow:
   - WBP 구조 리팩터
+  - 주간 계획 작성
